Test build-config forwarding through resolve_for_targets
TestBuildConfigPlumbing built its own coordinator, so it read back the
build_config it had just set rather than the one resolve_for_targets
forwards. Drive it through the real entry point instead, in
test_resolve.py alongside the other resolve_for_targets tests.
